backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
httpx>=0.24.0,<0.26.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
stripe==7.12.0
supabase==1.2.0
numpy==1.26.3
replicate==0.22.0
